requirements.txt
flask==1.1.2
psycopg2==2.8.5
discord.py[voice]==1.3.3
Werkzeug==1.0.1
boto3==1.12.39
flask-cors
numpy==1.18.4
requests==2.23.0